restructured files based on new models crate; added 'all' fundamentals to add crate
