+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cli;
+pub mod models;
 pub mod ops;
 pub mod utils;
